Reorganizing Files
Moved my email function and search files into separate folders
