+++ src/components/HomePage.js
@@ -22,9 +22,9 @@ export const HomePage = () => {
             </div>
 
             <div className="menu-bar">
-                <Link to="/admin-interface"><button>Admin Interface</button></Link>
-                <Link to="/send-email"><button>Send Email</button></Link>
-                <Link to="/search-menu"><button>Search Menu</button></Link>
+                <Link to="/admin-interface"><button className='menuBarButtons'>Admin Interface</button></Link>
+                <Link to="/send-email"><button className='menuBarButtons'>Send Email</button></Link>
+                <Link to="/search-menu"><button className='menuBarButtons'>Search Menu</button></Link>
             </div>
 
             <div className="blue-box">
